requirements.txt
numpy
scikit-build
scikit-learn
opencv-python
pandas
matplotlib
tensorboardX
tensorboard
yacs
tqdm
pickle-mixin
efficientnet_pytorch
h5py
timm
geffnet
torchgeometry
traitlets==5.1.1
pygments==2.4.1
lxml
scikit-image
webcolors
einops
bayesian-torch
